3.5.12 release (#231)
* Offline replay v1 (#216)

* First pass at implementing offline replay

* changes per review suggestions

* Fix tests by changing back defaultHost

* Format per suggestion (using ctrl+i in xcode)

* Bump version (#221)

* Fix timestamp on replays (#222)

* Bump version (#223)

* Skip logging of replay payloads (#224)

* Surface coordinates for geofences & places (#220)

* Include geometryCenter & geometryRadius in geofence object

* Include location in places object

* Rely on geometry object for radius and center

* Formatting

* RAD-5748 Anonymous config call in `trackOnce` (#219)

* add config call for anonymous before track

* pass track boolean

* switch to using usage

* resume usage

* set string const prior

* reverse args

* move anon usage config call deeper in the call stack

* more idiomatic ios signature

* bump version

* version bump

* 3.5.11 version bump

---------


* dont fire listeners for custom events (#218)

* Bump version to 3.5.12-beta.1 (#225)

* Set timeout to 25s for track/replay (#227)

* Bump to 3.5.12-beta.2

* Revert continuous preset to tracking none (#230)

* Include approachingThreshold in tripOptions (#214)

* Add tracking options

* Add condition to equals

* Only include in dictionaryValue if positive integer, imporve isEqual

* 3.5.12-beta.3

* Always clear last location if successful track (#232)

* Always clear last location if succesful track

* always clear ReplayBuffer upon successful request

---------


* serialize NSDates as long (#234)

* serialize NSDates as long

* dont cast to long

* dont cast to long

* not beta anymore

* not beta anymore

---------

+++ RadarSDK/Include/RadarTripOptions.h
@@ -57,7 +57,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic, assign) RadarRouteMode mode;
 
-@property (nonatomic, assign) UInt8 approachingThreshold;
+@property (nonatomic, assign) UInt16 approachingThreshold;
 
 + (RadarTripOptions *_Nullable)tripOptionsFromDictionary:(NSDictionary *)dict;
 - (NSDictionary *)dictionaryValue;
